jQuery goodies
Added support for detecting when the radial progress bar area is in the
viewport. Once it’s in the viewport the radial progress bars will fill
in to their assigned percentages.

jQuery code is in the _main.js file and will only run on the home page
to avoid unnecessary overhead.

If the “radial-progress” class is changed in the html the allMods
object in the jQuery code must be updated with the new class.